# restoWeb/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: RestoTouch
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /item:
     # binds a127 app logic to a route
    x-swagger-router-controller: item
    get:
      # used as the method name of the controller
      operationId: getAllItems
      description: Returns the list of items
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ItemsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: addItem
      description: Add a item
      parameters:
        - name: item
          description: item properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Item"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /item/{id}:
    x-swagger-router-controller: item
    get:
      description: Returns a specific item
      operationId: getItem
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Item"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a item
      # used as the method name of the controller
      operationId: updateItem
      parameters:
        - name: id
          type: number
          in: path
          required: true
        - name: category
          description: Category properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Item"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes an item
      # used as the method name of the controller
      operationId: deleteItem
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # -----------------------------------------/// Category ///
...
  /menu:
    # binds a127 app logic to a route
    x-swagger-router-controller: menu
    get:
      description: Returns the list of menus
      # used as the method name of the controller
      operationId: getAllMenu
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MenuListResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: saveMenu
      description: Add a menu
      # menu info to be stored
      parameters:
        - name: restaurant
          description: Menu properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Menu"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /menu/{menuId}:
    # binds a127 app logic to a route
    x-swagger-router-controller: menu
    get:
      description: Returns a specific menu
      # used as the method name of the controller
      operationId: getMenu
      parameters:
        - name: menuId
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Menu"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a menu
      # used as the method name of the controller
      operationId: updateMenu
      parameters:
        - name: menuId
          type: number
          in: path
          required: true
        - name: menu
          description: Menu properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Menu"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a menu
      # used as the method name of the controller
      operationId: delMenu
      parameters:
        - name: menuId
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /supportedLanguages:
    x-swagger-router-controller: account
    get:
      operationId: getSupportedLanguages
      description: Get the users supported languages
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SupportedLanguagesResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"         
  /accountSettings:
    x-swagger-router-controller: account
    put:
      operationId: saveAccountSettings
      description: Get the users account settings
      parameters:
        - name: accountSettings
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/AccountSettings"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: getAccountSettings
      description: Get the users account settings
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AccountSettings"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
    required:
      - success
      - description
...
  Category:
    type: object
    properties:
      supportedLanguages:
        type: array
        items:
          "$ref": "#/definitions/Language"
      translations:
        type: array
        items:
          "$ref": "#/definitions/CategoryTranslations"
      items:
        type: array
        items:
          "$ref": "#/definitions/Item"
  CategoryTranslations:
    type: object
    properties:
      name:
        type: string
      languageCode:
        type: string
    required:
      - name
      - languageCode
  ItemsResponse:
    required:
      - items
    properties:
      items:
        type: array
        items:
          "$ref": "#/definitions/Item"      
  Item:
    type: object
    properties:
      imageUrl:
        type: string
      itemSizes:
        type: array
        items:
          "$ref": "#/definitions/ItemSize"
      supportedLanguages:
        type: array
        items:
          "$ref": "#/definitions/Language"
      translations:
        type: array
        items:
          "$ref": "#/definitions/ItemTranslations"
      ingredientGroup:
        type: array
        items:
          "$ref": "#/definitions/IngredientGroup"
      categories:
        type: array
        items:
          "$ref": "#/definitions/Category"
    required:
      - imageUrl
  Ingredient:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
      allowQuantity:
        type: number
    required:
      - name
  IngredientGroup:
    type: object
    properties:
      name:
        type: string
      numberOfIngredientsAllowed:
        type: number
      allowQuantity:
        type: number
      ingredients:
        type: array
        items:
          "$ref": "#/definitions/Ingredient"
    required:
      - name
  ItemSize:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
    required:
      - name
      - price
  ItemTranslations:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      languageCode:
        type: string
    required:
      - name
      - description
      - languageCode
...
  Language:
    type: object
    properties:
      code:
        type: string
  SupportedLanguagesResponse:
    required:
      - supportedLanguages
    properties:
      supportedLanguages:
        type: array
        items:
          "$ref": "#/definitions/Language"
...
  MenuTranslations:
    type: object
    properties:
      name:
        type: string
      languageCode:
        type: string
    required:
      - name
      - languageCode
...
  AccountSettings:
    type: object
    properties:
      supportedLanguages:
        type: array
        items:
          "$ref": "#/definitions/Language"
...
  MenuListResponse:
    required:
      - menus
    properties:
      # The array of menus
      menus:
        type: array
        items:
          "$ref": "#/definitions/Menu"
  Menu:
    type: object
    properties:
      supportedLanguages:
        type: array
        items:
          "$ref": "#/definitions/Language"
      translations:
        type: array
        items:
          "$ref": "#/definitions/MenuTranslations"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string